Do not use network_mode: host for nginx
This is not really needed, nginx is always intended to be behind a
reverse proxy, so it can just expose its ports on the internal network.
This prevents having the 4808x ports exposed externally (though a user
can still expose them explicitly with a ports directive if needed).

This removes the "nginx" extra-hosts definition from the
bbb-graphql-middleware host container, since the "nginx" name now
resolves to the nginx internal IP normally.

In the haproxy config, this updates the IP address instead of using the
"nginx" name, as it seems that the haproxy resolv.conf does not list the
docker resolver but somehow uses the (host system) external DNS
directly.

Nginx also listens on port 8185. Previously this bound to 127.0.0.1 and
10.7.7.1 (the host-side IP of the internal network) explicitly,
presumably to prevent this port being available publically, but this is
no longer needed, so now it can just bind to the port directly.