Update network-config.md to current situation
This still listed old (now removed) containers and missed newly added
ones, which are now fixed.

This also makes the network_mode: host listings consistent, and removes
the "extra_hosts" listed for nginx (this does not really seem to have
much advantage in docs, unlike a list of allocated IP addresses which
*is* useful).

The list is slightly reordered to match the order in
docker-compose.tmpl.yml.