Add missing volume mounts for coturn and redis
These directories were marked as volumes in their dockerfile, but
nothing was explicitly mounted on them. This makes docker create an
unnamed volume and mount that when the container is created, which means
that on every down/up cycle, it would leave two unnamed volumes
lingering around.

Explicitly mounting these data directories prevents this.